I seem to be avoiding writing tests
