+++ Cargo.toml
@@ -49,3 +49,7 @@ clipboard = "0.5.0"
 # Terminal syntax highlighting
 syntect = "4.1.1"
 hex = "0.4.2"
+
+[dev-dependencies]
+assert_cmd = "0.10"
+predicates = "1"
