+++ src/errors.rs
@@ -23,9 +23,7 @@ pub enum LostTheWay {
     #[error("$HOME not set")]
     Homeless,
     /// Thrown when trying to load a theme which hasn't been added / doesn't exist
-    #[error(
-        "I don't have the {theme_name:?} theme. Add it from a theme file with the-way themes --add"
-    )]
+    #[error("Couldn't load theme {theme:?}. Are you sure it exists and is valid?")]
     ThemeError { theme: String },
     #[error("Couldn't copy to clipboard")]
     ClipboardError,
