+++ src/the_way/filter.rs
@@ -2,31 +2,47 @@
 use std::collections::HashSet;
 
 use anyhow::Error;
+use chrono::{Date, Utc};
+use structopt::StructOpt;
 
-use crate::the_way::cli::Filters;
-use crate::the_way::snippet::Snippet;
-use crate::the_way::TheWay;
+use crate::the_way::{snippet::Snippet, TheWay};
 use crate::utils;
+
+#[derive(StructOpt, Debug)]
+pub(crate) struct Filters {
+    /// Snippets from <DATE>
+    #[structopt(long, parse(try_from_str = utils::parse_date))]
+    pub(crate) from: Option<Date<Utc>>,
+    /// Snippets from <DATE>
+    #[structopt(long, parse(try_from_str = utils::parse_date))]
+    pub(crate) to: Option<Date<Utc>>,
+    /// Snippets written in <LANGUAGE> (multiple with 'lang1 lang2')
+    #[structopt(short, long)]
+    pub(crate) languages: Option<Vec<String>>,
+    /// Snippets with <TAG> (multiple with 'tag1 tag2')
+    #[structopt(short, long)]
+    pub(crate) tags: Option<Vec<String>>,
+}
 
 impl TheWay {
     /// Filters a list of snippets by given language/tag/date
     pub(crate) fn filter_snippets(&self, filters: &Filters) -> Result<Vec<Snippet>, Error> {
         let from_date = utils::date_start(filters.from);
         let to_date = utils::date_end(filters.to);
-        let snippets: Option<Vec<_>> = match filters.languages.clone() {
+        let snippets: Option<Vec<_>> = match &filters.languages {
             Some(languages) => Some(
                 self.get_snippets(
                     &languages
-                        .into_iter()
+                        .iter()
                         .flat_map(|language| {
-                            self.get_language_snippets(&language).unwrap_or_default()
+                            self.get_language_snippets(language).unwrap_or_default()
                         })
                         .collect::<Vec<_>>(),
                 )?,
             ),
             None => None,
         };
-        match (filters.tags.clone(), snippets) {
+        match (&filters.tags, snippets) {
             (Some(tags), Some(snippets)) => Ok(snippets
                 .into_iter()
                 .filter(|snippet| {
@@ -36,8 +52,8 @@ impl TheWay {
                 .collect()),
             (Some(tags), None) => {
                 let indices = tags
-                    .into_iter()
-                    .flat_map(|tag| self.get_tag_snippets(&tag).unwrap_or_default())
+                    .iter()
+                    .flat_map(|tag| self.get_tag_snippets(tag).unwrap_or_default())
                     .collect::<HashSet<_>>()
                     .into_iter()
                     .collect::<Vec<_>>();
